using long int

+++ 0x13-bit_manipulation/holberton.h
@@ -8,5 +8,5 @@ void print_binary(unsigned long int n);
 int get_bit(unsigned long int n, unsigned int index);
 int set_bit(unsigned long int *n, unsigned int index);
 int clear_bit(unsigned long int *n, unsigned int index);
-int _pow(unsigned int base, unsigned int exp);
+long int _pow(unsigned int base, unsigned int exp);
 #endif
